+++ examples/json.ts
@@ -1,26 +1,19 @@
 import type { JsonValue } from "emnorst";
-import { EOI, type Parser, choice, el, lazy, literal, many, qo, regex } from "../src";
-
-const sepBy = <T>(parser: Parser<T, string>, sep: Parser<unknown, string>) =>
-    qo<T[], string>(perform => {
-        const head = perform(parser);
-        const rest = perform(sep.then(parser).apply(many));
-        return [head, ...rest];
-    });
+import * as P from "parsea";
 
 // https://www.json.org/
 
-const ws = regex(/[ \n\r\t]*/);
+const ws = P.regex(/[ \n\r\t]*/);
 
-const jsonValue: Parser<JsonValue, string> = lazy(() =>
-    choice([
+export const json: P.Parser<JsonValue, string> = P.lazy(() =>
+    P.choice([
         object,
         array,
         string,
         number,
-        literal("true").return(true),
-        literal("false").return(false),
-        literal("null").return(null),
+        P.string("true").return(true),
+        P.string("false").return(false),
+        P.string("null").return(null),
     ]).between(ws),
 );
 
@@ -32,11 +25,12 @@ const escapeTable = {
     t: "\t",
 };
 
-const string = regex(/(?:\\(?:["\\/bfnrt]|u[0-9A-Fa-f]{4})|[^"\\])*/)
-    .between(el('"'))
+// cspell:ignore bfnrt
+const string = P.regex(/(?:\\(?:["\\/bfnrt]|u[0-9A-Fa-f]{4})|[^"\\])*/)
+    .between(P.el('"'))
     .map(escapedString =>
         escapedString.replace(/\\(u[0-9A-Fa-f]{4}|.)/g, (_, escape: string) => {
-            if (escape[0] === "u") {
+            if (escape.startsWith("u")) {
                 return String.fromCharCode(parseInt(escape.slice(1), 16));
             }
             if (escape in escapeTable) {
@@ -46,18 +40,52 @@ const string = regex(/(?:\\(?:["\\/bfnrt]|u[0-9A-Fa-f]{4})|[^"\\])*/)
         }),
     );
 
-const number = regex(/-?(0|[1-9]\d*)(.\d+)?([Ee][-+]?\d+)?/).map(Number);
+const number = P.regex(/-?(0|[1-9]\d*)(.\d+)?([Ee][-+]?\d+)?/).map(Number);
 
 const empty = ws.map<[]>(() => []);
 
-const array = jsonValue.apply(sepBy, el(",")).or(empty).between(el("["), el("]"));
+const array = P.sepBy(json, P.el(",")).skip(empty).between(P.el("["), P.el("]"));
 
-const keyValue = string.between(ws).skip(el(":")).and(jsonValue);
+const keyValue = P.seq([string.between(ws).skip(P.el(":")), json]);
 
-const object = keyValue
-    .apply(sepBy, el(","))
-    .or(empty)
-    .between(el("{"), el("}"))
-    .map<Record<string, JsonValue>>(Object.fromEntries);
+const object = P.sepBy(keyValue, P.el(","))
+    .skip(empty)
+    .between(P.el("{"), P.el("}"))
+    .map(Object.fromEntries<JsonValue>);
 
-export const jsonParser = jsonValue.skip(EOI);
+if (import.meta.vitest) {
+    const { test, expect } = import.meta.vitest;
+
+    test.each([
+        " null ",
+        " [ ] ",
+        " [ null , 0 ] ",
+        " {  } ",
+        ` { "1" : null , "2" : 0 } `,
+        "0",
+        "3.141592",
+        "4.2195e1",
+        "0.00E+00",
+        "-0e-0",
+    ])("%o", jsonString => {
+        expect(P.parseA(json, jsonString)).toEqual(JSON.parse(jsonString));
+    });
+
+    test.each(["00", "- 0", "0.", ".0"])("%o is invalid json.", n => {
+        expect(() => P.parseA(json, n)).toThrow(P.ParseAError);
+    });
+
+    test.each([
+        ['"'],
+        ["\\"],
+        ["/"],
+        ["b", "\b"],
+        ["f", "\f"],
+        ["n", "\n"],
+        ["r", "\r"],
+        ["t", "\t"],
+        ["u1234", "\u1234"],
+    ])("escape %s", (escapeChar, char = escapeChar) => {
+        expect(P.parseA(json, `"\\${escapeChar}"`)).toBe(char);
+    });
+}