+++ examples/s-expression.ts
@@ -1,16 +1,27 @@
-import { type Parser, choice, el, lazy, many, regex } from "../src";
+import { type Parser, choice, el, lazy, many, parseA, regex } from "parsea";
 
 export type SExpression = string | readonly SExpression[];
 
-const list = lazy(() => SExpression)
-    .apply(many)
-    .between(el("("), el(")"));
+const list = lazy(() => SExpression.apply(many).between(el("("), el(")")));
 
 export const SExpression: Parser<SExpression, string> = choice([
-    el("'")
-        .option()
-        .and(list)
-        .map(([quote, list]) => (quote ? ["quote", ...list] : list)),
+    el("'").then(list.map(list => ["quote", ...list])),
+    list,
     regex(/"([^"\\]|\\.)*"/),
     regex(/[^\s()"]+/),
 ]).between(regex(/\s*/));
+
+if (import.meta.vitest) {
+    const { test, expect } = import.meta.vitest;
+
+    test("Hello world!", () => {
+        expect(parseA(SExpression, '(print "Hello world!")')).toEqual([
+            "print",
+            '"Hello world!"',
+        ]);
+    });
+
+    test("quote list", () => {
+        expect(parseA(SExpression, "'(1 2 3 4)")).toEqual(["quote", "1", "2", "3", "4"]);
+    });
+}