+++ examples/script.test.ts
@@ -1,10 +1,10 @@
-import { describe, expect, test } from "@jest/globals";
-import { expr, stmt } from "./script";
+import { parseA } from "parsea";
+import { describe, expect, test } from "vitest";
+import { expression, statement } from "./script";
 
 describe("stmt", () => {
     test("Let", () => {
-        const result = stmt.parse("let hoge = 0;");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(statement, "let hoge = 0;")).toEqual({
             type: "Let",
             name: "hoge",
             init: { type: "Number", value: 0 },
@@ -12,44 +12,39 @@ describe("stmt", () => {
     });
 
     test("DefFn", () => {
-        const result = stmt.parse("fn main(arg) {};");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(statement, "fn main(arg) {};")).toEqual({
             type: "DefFn",
             name: "main",
             params: ["arg"],
-            body: { type: "Block", stmts: [], last: null },
+            body: { type: "Block", statements: [], last: null },
         });
     });
 
     test("Return", () => {
-        const result = stmt.parse("return;");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(statement, "return;")).toEqual({
             type: "Return",
             body: null,
         });
     });
 
     test("While", () => {
-        const result = stmt.parse("while (false) {};");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(statement, "while (false) {};")).toEqual({
             type: "While",
             test: { type: "Bool", value: false },
-            body: { type: "Block", stmts: [], last: null },
+            body: { type: "Block", statements: [], last: null },
         });
     });
 
     test("Break", () => {
-        const result = stmt.parse("break;");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(statement, "break;")).toEqual({
             type: "Break",
         });
     });
 
     test("Expr", () => {
-        const result = stmt.parse("0;");
-        expect(result.success && result.value).toEqual({
-            type: "Expr",
-            expr: { type: "Number", value: 0 },
+        expect(parseA(statement, "0;")).toEqual({
+            type: "Expression",
+            expression: { type: "Number", value: 0 },
         });
     });
 });
@@ -57,8 +52,7 @@ describe("stmt", () => {
 describe("expr", () => {
     describe("Bool", () => {
         test.each([true, false])("%s", value => {
-            const result = expr.parse(String(value));
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, String(value))).toEqual({
                 type: "Bool",
                 value,
             });
@@ -70,8 +64,7 @@ describe("expr", () => {
             ["int", "42", 42],
             ["minus float", "-273.25", -273.25],
         ])("%s (%f)", (_, src, value) => {
-            const result = expr.parse(src);
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, src)).toEqual({
                 type: "Number",
                 value,
             });
@@ -79,16 +72,14 @@ describe("expr", () => {
     });
 
     test("String", () => {
-        const result = expr.parse(`"Hello world!"`);
-        expect(result.success && result.value).toEqual({
+        expect(parseA(expression, `"Hello world!"`)).toEqual({
             type: "String",
             value: "Hello world!",
         });
     });
 
     test("Ident", () => {
-        const result = expr.parse("hoge");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(expression, "hoge")).toEqual({
             type: "Ident",
             name: "hoge",
         });
@@ -96,15 +87,13 @@ describe("expr", () => {
 
     describe("Tuple", () => {
         test("empty", () => {
-            const result = expr.parse("()");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "()")).toEqual({
                 type: "Tuple",
                 elements: [],
             });
         });
         test("elements", () => {
-            const result = expr.parse(`(0, "", hoge)`);
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, `(0, "", hoge)`)).toEqual({
                 type: "Tuple",
                 elements: [
                     { type: "Number", value: 0 },
@@ -114,8 +103,7 @@ describe("expr", () => {
             });
         });
         test("trailing comma", () => {
-            const result = expr.parse("(0, )");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "(0, )")).toEqual({
                 type: "Tuple",
                 elements: [{ type: "Number", value: 0 }],
             });
@@ -124,26 +112,25 @@ describe("expr", () => {
 
     describe("Block", () => {
         test("empty", () => {
-            const result = expr.parse("{}");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "{}")).toEqual({
                 type: "Block",
-                stmts: [],
+                statements: [],
                 last: null,
             });
         });
         test("stmts", () => {
-            const result = expr.parse("{ 0; }");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "{ 0; }")).toEqual({
                 type: "Block",
-                stmts: [{ type: "Expr", expr: { type: "Number", value: 0 } }],
+                statements: [
+                    { type: "Expression", expression: { type: "Number", value: 0 } },
+                ],
                 last: null,
             });
         });
         test("expr", () => {
-            const result = expr.parse("{ 0 }");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "{ 0 }")).toEqual({
                 type: "Block",
-                stmts: [],
+                statements: [],
                 last: { type: "Number", value: 0 },
             });
         });
@@ -151,17 +138,15 @@ describe("expr", () => {
 
     describe("If", () => {
         test("then only", () => {
-            const result = expr.parse("if (true) {}");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "if (true) {}")).toEqual({
                 type: "If",
                 test: { type: "Bool", value: true },
-                then: { type: "Block", stmts: [], last: null },
+                then: { type: "Block", statements: [], last: null },
                 else: null,
             });
         });
         test("with else", () => {
-            const result = expr.parse("if (true) 1 else 0");
-            expect(result.success && result.value).toEqual({
+            expect(parseA(expression, "if (true) 1 else 0")).toEqual({
                 type: "If",
                 test: { type: "Bool", value: true },
                 then: { type: "Number", value: 1 },
@@ -171,8 +156,7 @@ describe("expr", () => {
     });
 
     test("tail", () => {
-        const result = expr.parse("f(0)(1)");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(expression, "f(0)(1)")).toEqual({
             type: "Call",
             callee: {
                 type: "Call",
@@ -184,8 +168,7 @@ describe("expr", () => {
     });
 
     test("Call", () => {
-        const result = expr.parse(`print("Hello world!")`);
-        expect(result.success && result.value).toEqual({
+        expect(parseA(expression, `print("Hello world!")`)).toEqual({
             type: "Call",
             callee: { type: "Ident", name: "print" },
             arguments: [{ type: "String", value: "Hello world!" }],
@@ -193,8 +176,7 @@ describe("expr", () => {
     });
 
     test("Property", () => {
-        const result = expr.parse("().property");
-        expect(result.success && result.value).toEqual({
+        expect(parseA(expression, "().property")).toEqual({
             type: "Property",
             target: { type: "Tuple", elements: [] },
             name: "property",
@@ -203,11 +185,11 @@ describe("expr", () => {
 });
 
 test("script", () => {
-    const result = expr.parse(`{
+    const source = `{
         fn main() {
             print("Hello world!");
             ()
         };
-    }`);
-    expect(result.success && result.value).toMatchSnapshot();
+    }`;
+    expect(parseA(expression, source)).toMatchSnapshot();
 });