+++ examples/script.ts
@@ -1,66 +1,64 @@
-import * as P from "../src";
+import * as P from "parsea";
 
-export type Expr =
+export type Expression =
     | { type: "Bool"; value: boolean }
     | { type: "Number"; value: number }
     | { type: "String"; value: string }
-    | { type: "Tuple"; elements: readonly Expr[] }
-    | { type: "Block"; stmts: Stmt[]; last: Expr | null }
-    | { type: "If"; test: Expr; then: Expr; else: Expr | null }
+    | { type: "Tuple"; elements: readonly Expression[] }
+    | { type: "Block"; statements: Statement[]; last: Expression | null }
+    | { type: "If"; test: Expression; then: Expression; else: Expression | null }
     | { type: "Ident"; name: string }
-    | { type: "Call"; callee: Expr; arguments: readonly Expr[] }
-    | { type: "Property"; target: Expr; name: string };
+    | { type: "Call"; callee: Expression; arguments: readonly Expression[] }
+    | { type: "Property"; target: Expression; name: string };
 
-export const expr: P.Parser<Expr, string> = P.lazy(() =>
+export const expression: P.Parser<Expression, string> = P.lazy(() =>
     P.choice([Bool, Number, String, Tuple, Block, If, Ident]).between(ws).flatMap(tail),
 );
 
-const sepBy = <T, S>(
-    parser: P.Parser<T, S>,
-    sep: P.Parser<unknown, S>,
-): P.Parser<T[], S> => {
-    return P.qo(perform => {
-        const xs: T[] = [];
-        perform.try(undefined, () => {
-            for (;;) {
-                xs.push(perform(parser, { allowPartial: true }));
-                perform(sep, { allowPartial: true });
-            }
-        });
-        return xs;
-    });
-};
-
 const ws = P.regex(/\s*/);
 
-const keyword = (keyword: string): P.Parser<unknown, string> => {
-    return P.literal(keyword).then(P.notFollowedBy(P.regex(/\w/)));
+const keywords = [
+    "let",
+    "fn",
+    "return",
+    "while",
+    "break",
+    "true",
+    "false",
+    "if",
+    "else",
+] as const;
+
+const keyword = (keyword: (typeof keywords)[number]): P.Parser<unknown, string> => {
+    return P.literal(keyword).then(P.regex(/\b/));
 };
 
-const Ident = P.regex(/\w+/).map(name => ({ type: "Ident", name }) satisfies Expr);
+const Ident = P.notFollowedBy(P.choice(keywords.map(keyword)))
+    .then(P.regex(/\b\w+\b/))
+    .map(name => ({ type: "Ident", name }) satisfies Expression);
 
-export type Stmt =
-    | { type: "Let"; name: string; init: Expr }
-    | { type: "DefFn"; name: string; params: readonly string[]; body: Expr }
-    | { type: "Return"; body: Expr | null }
-    | { type: "While"; test: Expr; body: Expr }
+export type Statement =
+    | { type: "Let"; name: string; init: Expression }
+    | { type: "DefFn"; name: string; params: readonly string[]; body: Expression }
+    | { type: "Return"; body: Expression | null }
+    | { type: "While"; test: Expression; body: Expression }
     | { type: "Break" }
-    | { type: "Expr"; expr: Expr };
+    | { type: "Expression"; expression: Expression };
 
-const Let = keyword("let")
-    .then(Ident.between(ws))
-    .skip(P.el("="))
-    .andMap(expr, ({ name }, init): Stmt => ({ type: "Let", name, init }));
+const Let = P.seq([
+    keyword("let").then(Ident.between(ws)).skip(P.el("=")),
+    expression,
+]).map(([{ name }, init]): Statement => ({ type: "Let", name, init }));
 
 const DefFn = P.seq([
     keyword("fn").then(Ident.between(ws)),
     Ident.between(ws)
-        .apply(sepBy, P.el(","))
+        .apply(P.sepBy, P.el(","), { trailing: "allow" })
         .skip(ws)
         .between(P.el("("), P.el(")"))
         .map(nodes => nodes.map(node => node.name)),
-    expr,
-]).map<Stmt>(([{ name }, params, body]) => ({
+    expression,
+]).map<Statement>(([{ name }, params, body]) => ({
     type: "DefFn",
     name,
     params,
@@ -68,26 +66,31 @@ const DefFn = P.seq([
 }));
 
 const Return = keyword("return")
-    .then(expr.option(null))
+    .then(expression.option(null))
     .skip(ws)
-    .map<Stmt>(body => ({ type: "Return", body }));
+    .map<Statement>(body => ({ type: "Return", body }));
 
-const While = keyword("while")
-    .skip(ws)
-    .then(expr.between(P.el("("), P.el(")")))
-    .andMap(expr, (test, body): Stmt => ({ type: "While", test, body }));
+const While = P.seq([
+    keyword("while")
+        .skip(ws)
+        .then(expression.between(P.el("("), P.el(")"))),
+    expression,
+]).map(([test, body]): Statement => ({ type: "While", test, body }));
 
-const Break = keyword("break").return<Stmt>({ type: "Break" }).skip(ws);
+const Break = keyword("break").return<Statement>({ type: "Break" }).skip(ws);
 
-const Expr = expr.map<Stmt>(expr => ({ type: "Expr", expr }));
+const Expression = expression.map<Statement>(expression => ({
+    type: "Expression",
+    expression,
+}));
 
-export const stmt: P.Parser<Stmt, string> = P.choice([
+export const statement: P.Parser<Statement, string> = P.choice([
     Let,
     DefFn,
     Return,
     While,
     Break,
-    Expr,
+    Expression,
 ])
     .skip(P.el(";"))
     .between(ws);
@@ -95,41 +98,35 @@ export const stmt: P.Parser<Stmt, string> = P.choice([
 const Bool = P.choice([
     keyword("true").return(true),
     keyword("false").return(false),
-]).map<Expr>(value => ({ type: "Bool", value }));
+]).map<Expression>(value => ({ type: "Bool", value }));
 
 const digit = P.oneOf("0123456789");
-const digits = digit.apply(
-    P.manyAccum<string, string, string>,
-    (accum, digit) => accum + digit,
-    () => "",
-    { min: 1 },
-);
+const digits = digit.apply(P.many, { min: 1 }).map(digits => digits.join(""));
 
-const sign = P.oneOf(["+", "-"] as const).option("+" as const);
+const sign = P.oneOf(["+", "-"]).option("+");
 
-const Number = sign
-    .andMap(digits, (sign, digits) => sign + digits)
-    .andMap(P.el(".").then(digits).option(""), (int, fDigits) => {
+const Number = P.seq([sign, digits, P.el(".").then(digits).option("")]).map(
+    ([sign, intDigits, floatDigits]) => {
         return {
             type: "Number",
-            value: parseFloat(`${int}.${fDigits}`),
-        } satisfies Expr;
-    });
+            value: parseFloat(`${sign}${intDigits}.${floatDigits}`),
+        } satisfies Expression;
+    },
+);
 
 const String = P.regex(/([^"\\]|\\.)*/)
     .between(P.el('"'))
-    .map<Expr>(value => ({ type: "String", value }));
+    .map<Expression>(value => ({ type: "String", value }));
 
-const Tuple = expr
-    .apply(sepBy, P.el(","))
+const Tuple = expression
+    .apply(P.sepBy, P.el(","), { trailing: "allow" })
     .skip(ws)
     .between(P.el("("), P.el(")"))
-    .map(elements => ({ type: "Tuple", elements }) satisfies Expr);
+    .map(elements => ({ type: "Tuple", elements }) satisfies Expression);
 
-const Block = stmt
-    .apply(P.many)
-    .andMap(expr.option(null), (stmts, last) => {
-        return { type: "Block", stmts, last } satisfies Expr;
+const Block = P.seq([statement.apply(P.many), expression.option(null)])
+    .map(([statements, last]) => {
+        return { type: "Block", statements, last } satisfies Expression;
     })
     .skip(ws)
     .between(P.el("{"), P.el("}"));
@@ -137,26 +134,23 @@ const Block = stmt
 const If = P.seq([
     keyword("if")
         .then(ws)
-        .then(expr.between(P.el("("), P.el(")"))),
-    expr,
-    keyword("else").then(expr).option(null),
-]).map<Expr>(([test, then, else_]) => ({
+        .then(expression.between(P.el("("), P.el(")"))),
+    expression,
+    keyword("else").then(expression).option(null),
+]).map<Expression>(([test, then, else_]) => ({
     type: "If",
     test,
     then,
     else: else_,
 }));
 
-const tail = (expr: Expr) =>
+const tail = (target: Expression) =>
     P.choice([Call, Property])
         .skip(ws)
-        .apply(
-            P.manyAccum<(callee: Expr) => Expr, Expr, string>,
-            (expr, tail) => tail(expr),
-            () => expr,
-        );
+        .apply(P.many)
+        .map(tails => tails.reduce((expr, tail) => tail(expr), target));
 
-const Call = Tuple.map<(callee: Expr) => Expr>(({ elements }) => callee => ({
+const Call = Tuple.map<(callee: Expression) => Expression>(({ elements }) => callee => ({
     type: "Call",
     callee,
     arguments: elements,
@@ -165,7 +159,7 @@ const Call = Tuple.map<(callee: Expr) => Expr>(({ elements }) => callee => ({
 const Property = P.el(".")
     .then(ws)
     .then(Ident)
-    .map<(target: Expr) => Expr>(({ name }) => target => ({
+    .map<(target: Expression) => Expression>(({ name }) => target => ({
         type: "Property",
         target,
         name,