+++ src/context.test.ts
@@ -1,10 +1,8 @@
-import { describe, expect, test } from "@jest/globals";
+import { describe, expect, test } from "vitest";
 import { Context } from "./context";
 
 describe("Context", () => {
     test("sourceにArrayLikeでない値を入れるとエラー", () => {
-        expect(() => {
-            new Context({ length: NaN }, {});
-        }).toThrow(TypeError);
+        expect(() => new Context({ length: NaN }, {})).toThrow(TypeError);
     });
 });