+++ src/do.ts
@@ -1,9 +1,16 @@
 import { has } from "emnorst";
 import type { Config } from "./context";
 import { Parser } from "./parser";
-import { updateState } from "./state";
+import { type ParseState, updateState } from "./state";
 
 const doError = /* #__PURE__ */ Symbol("parsea.doError");
+
+const get = (error: unknown) => {
+    if (has(error, doError)) {
+        return error[doError] as PerformOptions | undefined;
+    }
+    throw error;
+};
 
 export type PerformOptions = {
     allowPartial?: boolean;
@@ -11,7 +18,12 @@ export type PerformOptions = {
 
 export type Perform<S> = {
     <T>(parser: Parser<T, S>, options?: PerformOptions): T;
-    try<T, U = T>(defaultValue: T, runner: () => U): T | U;
+    option<T>(parser: Parser<T, S>): T | undefined;
+    option<T, const U = T>(parser: Parser<T, S>, defaultValue: U): T | U;
+    try<const T>(runner: () => T): T | undefined;
+    try<const T, const U = T>(runner: () => T, defaultValue: U): T | U;
+    many<T>(parser: Parser<T, S>): T[];
+    while(runner: () => void): void;
 };
 
 export const qo = <T, S>(
@@ -20,32 +32,57 @@ export const qo = <T, S>(
     new Parser((state, context) => {
         const perform: Perform<S> = (parser, options) => {
             const newState = parser.run(state, context);
-            if (newState == null) {
-                throw { [doError]: options };
-            }
+            if (newState == null) throw { [doError]: options };
             return (state = newState).v;
         };
 
-        perform.try = (defaultValue, runner) => {
+        perform.option = ((parser, defaultValue) => {
+            const newState = parser.run(state, context);
+            return newState == null ? defaultValue : (state = newState).v;
+        }) as Perform<S>["option"];
+
+        perform.try = ((runner, defaultValue) => {
             const beforeTryState = state;
             try {
                 return runner();
             } catch (error) {
-                if (!has(error, doError)) {
-                    throw error;
-                }
-                const options = error[doError] as PerformOptions | undefined;
+                const options = get(error);
                 if (!options?.allowPartial) {
                     state = beforeTryState;
                 }
                 return defaultValue;
             }
+        }) as Perform<S>["try"];
+
+        perform.many = parser => {
+            const result = [];
+            while (true) {
+                const newState = parser.run(state, context);
+                if (newState == null || !(state.i < newState.i)) break;
+                result.push((state = newState).v);
+            }
+            return result;
         };
 
-        return perform.try(null, () => {
+        perform.while = runner => {
+            let beforeWhileState!: ParseState<unknown>;
+            try {
+                do {
+                    beforeWhileState = state;
+                    runner();
+                } while (beforeWhileState.i < state.i);
+            } catch (error) {
+                const options = get(error);
+                if (!(options?.allowPartial ?? true)) {
+                    state = beforeWhileState;
+                }
+            }
+        };
+
+        return perform.try(() => {
             const value = runner(perform, context.cfg);
             return updateState(state, value);
-        });
+        }, null);
     });
 
 export { qo as do_ };