+++ src/parser.test.ts
@@ -1,5 +1,6 @@
-import { describe, expect, jest, test } from "@jest/globals";
+import { describe, expect, test, vi } from "vitest";
 import { label } from "./error";
+import { parseA } from "./parsea";
 import type { Parser } from "./parser";
 import { fail, literal, pure } from "./primitive";
 
@@ -13,32 +14,25 @@ describe("Parser", () => {
     });
 
     test("map", () => {
-        const fn = jest.fn<() => string>().mockReturnValue("hoge");
-        const result = pure(null).map(fn).parse([]);
-        expect(fn).lastCalledWith(null, {});
-        expect(result).toHaveProperty("value", "hoge");
+        const fn = vi.fn<() => string>().mockReturnValue("hoge");
+
+        expect(parseA(pure("fuga").map(fn), [])).toBe("hoge");
+        expect(fn).lastCalledWith("fuga", {});
     });
 
     test("flatMap", () => {
-        const fn = jest.fn<() => Parser<string>>().mockReturnValue(pure("hoge"));
-        const result = pure(null).flatMap(fn).parse([]);
-        expect(fn).lastCalledWith(null, {});
-        expect(result).toHaveProperty("value", "hoge");
+        const fn = vi.fn<() => Parser<string>>().mockReturnValue(pure("hoge"));
+
+        expect(parseA(pure("fuga").flatMap(fn), [])).toBe("hoge");
+        expect(fn).lastCalledWith("fuga", {});
     });
 
     describe("option", () => {
         test("success", () => {
-            expect(fail().option("some default value").parse([])).toMatchObject({
-                success: true,
-                value: "some default value",
-            });
+            expect(parseA(fail().option("defaultValue"), [])).toBe("defaultValue");
         });
-
         test("default", () => {
-            expect(fail().option().parse([])).toMatchObject({
-                success: true,
-                value: undefined,
-            });
+            expect(parseA(fail().option(), [])).toBeUndefined();
         });
     });
 });