+++ src/primitive.test.ts
@@ -1,78 +1,70 @@
-import { describe, expect, jest, test } from "@jest/globals";
-import { expected } from "./error";
-import { ANY_EL, EOI, el, literal, pure, satisfy } from "./primitive";
+import { describe, expect, test, vi } from "vitest";
+import { ParseAError, parseA } from "./parsea";
+import { anyEl, el, eoi, literal, pure, satisfy } from "./primitive";
 
 test("pure", () => {
-    const symbol = Symbol("ID");
-    expect(pure(symbol).parse([])).toHaveProperty("value", symbol);
+    const symbol = Symbol();
+    expect(parseA(pure(symbol), [])).toBe(symbol);
 });
 
-describe("EOI", () => {
+describe("eoi", () => {
     test("end of input", () => {
-        expect(ANY_EL.then(EOI).parse(["el"])).toHaveProperty("success", true);
+        expect(parseA(anyEl().then(eoi), ["el"])).toBe("el");
     });
     test("消費しきっていない要素がある場合失敗", () => {
-        expect(EOI.parse(["el"])).toHaveProperty("success", false);
+        expect(() => parseA(eoi, ["el"])).toThrow();
     });
 });
 
-describe("ANY_EL", () => {
+describe("anyEl", () => {
     test("lengthが1以上必要", () => {
-        expect(ANY_EL.parse([])).toHaveProperty("success", false);
+        expect(() => parseA(anyEl(), [])).toThrow();
     });
     test("長さを1消費", () => {
-        expect(ANY_EL.parse([0])).toHaveProperty("index", 1);
+        expect(() => parseA(anyEl().skip(anyEl()), [0])).toThrow();
     });
     test("任意の要素で成功する", () => {
-        expect(ANY_EL.parse(["el"])).toHaveProperty("value", "el");
+        expect(parseA(anyEl(), ["el"])).toBe("el");
     });
 });
 
 describe("el", () => {
-    test("SameValueZeroで判定", () => {
-        expect(el(1).parse([1])).toHaveProperty("value", 1);
-        expect(el(2).parse([1])).toHaveProperty("success", false);
-        expect(el(NaN).parse([NaN])).toHaveProperty("value", NaN);
-        expect(el(-0).parse([0])).toHaveProperty("value", 0);
+    test("SameValueZero", () => {
+        expect(parseA(el(1), [1])).toBe(1);
+        expect(() => parseA(el(2), [1])).toThrow();
+        expect(parseA(el(NaN), [NaN])).toBe(NaN);
+        expect(parseA(el(-0), [0])).toBe(0);
     });
 });
 
 describe("satisfy", () => {
-    test("残りの長さが1以上必要", () => {
-        const mock = jest.fn<() => boolean>().mockReturnValue(true);
-        expect(satisfy(mock).parse([])).toHaveProperty("success", false);
-        expect(mock).toHaveBeenCalledTimes(0);
+    test("success", () => {
+        const parser = satisfy(() => true);
+        expect(parseA(parser, [1])).toBe(1);
     });
-    test("条件に合う要素で成功", () => {
-        const EvenNumberParser = satisfy(el => (el as number) % 2 === 0);
-        expect(EvenNumberParser.parse([8])).toHaveProperty("success", true);
-        expect(EvenNumberParser.parse([7])).toHaveProperty("success", false);
+    test("failure", () => {
+        const parser = satisfy(() => false);
+        expect(() => parseA(parser, [1])).toThrow(ParseAError);
     });
-    test("valueは要素", () => {
-        expect(satisfy(() => true).parse([6])).toHaveProperty("value", 6);
+    test("empty", () => {
+        const fn = vi.fn<() => boolean>();
+        expect(() => parseA(satisfy(fn), [])).toThrow(ParseAError);
+        expect(fn).toHaveBeenCalledTimes(0);
     });
 });
 
 describe("literal", () => {
-    test("sourceが短いと失敗", () => {
-        expect(literal("abc").parse("ab")).toEqual({
-            success: false,
-            index: 0,
-            errors: [expected("abc")],
-        });
+    test("empty", () => {
+        expect(parseA(literal([]), [])).toEqual([]);
     });
-    test("空", () => {
-        expect(literal([]).parse([])).toHaveProperty("value", []);
+    test("sourceが短いと失敗", () => {
+        expect(() => parseA(literal("abc"), "ab")).toThrow(ParseAError);
     });
     test("違う要素で失敗", () => {
-        expect(literal("ふんいき").parse("ふいんき")).toEqual({
-            success: false,
-            index: 0,
-            errors: [expected("ふんいき")],
-        });
+        expect(() => parseA(literal("ふんいき"), "ふいんき")).toThrow(ParseAError);
     });
-    test("SameValueZeroで判定", () => {
+    test("SameValueZero", () => {
         const parser = literal(["hoge", NaN, -0]);
-        expect(parser.parse(["hoge", NaN, 0])).toHaveProperty("value", ["hoge", NaN, -0]);
+        expect(parseA(parser, ["hoge", NaN, 0])).toEqual(["hoge", NaN, -0]);
     });
 });